Add more activiy we need.

+++ Rapeprevention/app/src/main/java/hackmate/rapeprevention/Models/Model.java
@@ -2,12 +2,11 @@ package hackmate.rapeprevention.Models;
 
 import android.location.Location;
 import java.util.ArrayList;
-import java.util.List;
 
 public class Model {
   public Observable<Location> userLocation = new Observable<>();
   public Observable<String> userName = new Observable<>();
-  public Observable<ArrayList<Integer>> reactionTime = Observable.from(new ArrayList<Integer>());
+  public Observable<ArrayList<Long>> reactionTime = Observable.from(new ArrayList<Long>());
 
   private Model() {}
   private static Model ONLY = new Model();
